requirements.txt
fastapi
uvicorn
torch
transformers
onnxruntime
pdfplumber
easyocr
pdf2image
pillow
numpy
